refactor(prism): rewrite as a plugin
